accessories

+++ Assets/Scripts/CharacterController2D.cs
@@ -62,7 +62,7 @@ public class CharacterController2D : MonoBehaviour
     public UnityEvent OnLandEvent;
     public UnityEvent OnHeavyLandEvent;
 
-    public void setCCStats(float jfm, float hms, int mj, int ac){
+    public void setCCStats(float jfm, float hms, int mj, float ac){
             m_JumpForce = m_JumpForce * jfm;
             horizontalMovementSpeed =  horizontalMovementSpeed * hms;
             maxJumps = maxJumps + mj;
